Reorganize database code some more
In order for globalSetup and database-testing to share code without setting a bunch of extra connections, I had to create a third module for the more abstract stuff.
